Replacing the message builder.
I realised in a sleepy state this morning that I was thinking about the message builder wrong. Rather than storing a lot of slightly dissimilar logic twice for displaying a message, we now create the messages using a HammerTextBuilder and translate that in the new IMessageSender class. This allows us to define formatting once, rather than twice.

Big thanks to the Sponge guys for the inspiration for this.